adding find libs

+++ find_libs.py
@@ -13,8 +13,11 @@ def recursive_find(base, pattern="*.so"):
 
 def main(src, dest):
     for lib in recursive_find(src, "*.so"):
+        lib = os.path.realpath(lib)
         lib_dir = os.path.dirname(src).replace(src, "").strip("/")
         dest_lib = os.path.join(dest, lib_dir, os.path.basename(lib))
+        if os.path.exists(dest_lib):
+            continue
         print("Copying %s to %s" % (lib, dest_lib))
         shutil.copyfile(lib, dest_lib)
